Use new Assembler as disassembler, replacing the old Parser

+++ src/assembler/program.h
@@ -43,7 +43,7 @@ struct ProgramSection {
 struct Program {
     unsigned long entryPoint = 0;
     std::map<QString, ProgramSection> sections;
-    std::map<unsigned long, QString> symbols;
+    std::map<uint32_t, QString> symbols;
 
     const ProgramSection* getSection(const QString& name) const {
         const auto secIter =
@@ -56,4 +56,5 @@ struct Program {
         return &secIter->second;
     }
 };
+
 }  // namespace Ripes
